Extract shared klog init with newline-escaping stderr writer
Move `newlineEscapingStderrWriter` and `klog` initialization into `pkg/util/log` so both `aws-s3-csi-driver` and `aws-s3-csi-mounter` use the same logging setup writing to stderr with escaped newlines.
